Rename ConsensusCommitWithGrpcStorageIntegrationTest
